requirements.txt
PyQt6>=6.9.1
pynput>=1.7.8
mss>=10.0.0
requests>=2.32.4
Pillow>=11.2.1
betterproto>=2.0.0b7
keyboard~=0.13.5; sys_platform == 'win32'
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'
python-xlib~=0.33; sys_platform == 'linux'
websockets>=15.0.1
meikiocr>=0.3.1
onnxruntime==1.20.1; sys_platform == 'win32' # newer versions cause issues on windows
PyGObject==3.56.1; sys_platform == 'linux'
